+++ src/ai/behavior_tree.hpp
@@ -2,11 +2,12 @@
 #define BEHAVIOR_TREE_H
 
 #include <ai/behavior.hpp>
+#include <std_includes.hpp>
 
 class BehaviorTree {
 private:
   ABehavior* _root;
-  float _frame;
+  uint8_t _frame;
 public:
   BehaviorTree();
   ~BehaviorTree();
@@ -19,7 +20,7 @@ public:
      game frame rate. Example: a tick_rate of 5 means it ticks every 5
      frame.
   */
-  void bt_update(float tick_rate);
+  void bt_update(uint8_t tick_rate);
   void dealloc_root();
 };
